+++ projects/angular-material-extensions/select-country/src/lib/i18n/index.ts
@@ -1,4 +1,5 @@
 export * from './br';
+export * from './by';
 export * from './de';
 export * from './en';
 export * from './es';
@@ -10,3 +11,5 @@ export * from './ca'; //Catalan
 export * from './eu'; //Basque
 export * from './nl';
 export * from './pt';
+export * from './ru';
+export * from './ua';